Moved the UpdateTerminalListener to the terminals package.
